Initial render settings now stick on startup.
