+++ Assets/Source/Application/Bridge/EditorSettings.cs.meta
